requirements.txt
Django==3.2
djangorestframework==3.11.1
pika==1.3.1
coverage==7.2.7
